Add files via upload
You can find the Work in progress part at the bottom of the code. I created a modified version of the Profit_top_k function and it's called Profit_top_k_1. The reason being to integrage it in crossvalidation... check it out, if I find a way out I will update