Added an attempt at the building the scoring function.
Made a scoring function for the top 20 profit. It is not 100% perfect yet but I think with some improvements we can get there.
Also added an example of how the parameter tuning needs to go using cross validation and the scoring function. Use this when tuning all other models with the improved scoring function.