this is a file with full models and final result
now i am suspicious that we did not take cost into consideration. so currently we captured many cost = 0.1 cases so we get the low score with Profit_top_20. Because if we take a look at the AUC our score is quite good (0.93+). In our function we only care about probability as well.