Added the first version of the graph
This graph shows two communnities: the parties and Flanders and Wallonia
Additionally, labels with the parties are added to the party noded. Vlaams Belang and NVA are very close to each other. PVDA is in the middle between Flanders and Wallonia. This makes sense as the party is present in Flanders and Wallonia